🔀 :: (#55) - 내 주문 가져오기

+++ app/src/main/java/com/g3c1/oasis_android/feature_seat/domain/repository_impl/SeatRepositoryImpl.kt
@@ -1,7 +1,7 @@
 package com.g3c1.oasis_android.feature_seat.domain.repository_impl
 
-import com.g3c1.oasis_android.feature_seat.data.dto.SeatDTO
 import com.g3c1.oasis_android.feature_seat.data.data_soure.SeatDataSource
+import com.g3c1.oasis_android.feature_seat.data.dto.SeatDTO
 import com.g3c1.oasis_android.feature_seat.domain.repository.SeatRepository
 import com.g3c1.oasis_android.remote.util.ApiState
 import kotlinx.coroutines.flow.Flow
@@ -9,13 +9,13 @@ import javax.inject.Inject
 
 class SeatRepositoryImpl @Inject constructor(
     private val dataSource: SeatDataSource
-): SeatRepository {
+) : SeatRepository {
 
     override suspend fun getSeatDataRepository(): Flow<ApiState<List<SeatDTO>>> {
         return dataSource.getSeatInfo()
     }
 
-    override suspend fun patchSeatDataRepository(seatId: Int): Flow<ApiState<Unit>> {
+    override suspend fun patchSeatDataRepository(seatId: Long): Flow<ApiState<Unit>> {
         return dataSource.patchSeatData(seatId)
     }
 }
