+++ app/src/main/java/com/g3c1/oasis_android/feature_seat/domain/use_case/PatchSeatDataUseCase.kt
@@ -6,5 +6,5 @@ import javax.inject.Inject
 class PatchSeatDataUseCase @Inject constructor(
     private val repository: SeatRepository
 ){
-    suspend fun patchSeatDataUseCase(seatId: Int) = repository.patchSeatDataRepository(seatId)
+    suspend fun patchSeatDataUseCase(seatId: Long) = repository.patchSeatDataRepository(seatId)
 }
